refactor: isolate archive runtime and publication helpers
